Add suggested changes
