+++ grr/server/grr_response_server/gui/ui/components/client_details/client_details.scss
@@ -35,8 +35,6 @@
 
       .timestamp {
         margin-left: 20px;
-        font-family: Roboto;
-        font-style: normal;
         font-weight: 500;
         font-size: 11px;
         line-height: 16px;
