+++ grr/server/grr_response_server/gui/ui/lib/api/api_interfaces.ts
@@ -171,7 +171,7 @@ export declare interface ApiClient {
  * ApiGetClientVersionsResult proto mapping.
  */
 export declare interface ApiGetClientVersionsResult {
-  readonly items: ReadonlyArray<ApiClient>;
+  readonly items?: ReadonlyArray<ApiClient>;
 }
 
 /**
